requirements.txt
streamlit
langchain
langchain-google-genai
langchain-core
langchain-community
langchain-chroma
chromadb
python-dotenv
pypdf